Step6: 타워판매 까지 , 대포종류도 바꾸

+++ mythird_game/Assets/Scripts/Projectile.cs
@@ -6,9 +6,9 @@ public class Projectile : MonoBehaviour
     private Movement2D movement2D;
 
     private Transform target;
-    private int damage;
+    private float damage;
 
-    public void Setup(Transform target, int damage)
+    public void Setup(Transform target, float damage)
     {
         movement2D = GetComponent<Movement2D>();
         this.target = target;
